refactor: organize services into modules
