+++ frontend/src/app/core/services/api/practice.service.ts
@@ -5,7 +5,7 @@ import {
   ActiveSession,
   CardResponse,
   SessionSummary,
-} from '../../state/practice-session/models/practice-session.model';
+} from '@app/state/practice-session/models/practice-session.model';
 
 @Injectable({ providedIn: 'root' })
 export class PracticeService {
