+++ frontend/src/app/core/services/config/config.service.ts
