+++ frontend/src/app/core/services/utils/flow-parsing.service.ts
