+++ frontend/src/app/features/memorize/decks/components/create-deck-modal/create-deck-modal.component.ts
@@ -2,7 +2,7 @@
 import { Component, Output, EventEmitter, Input, OnInit } from '@angular/core';
 import { CommonModule } from '@angular/common';
 import { FormsModule } from '@angular/forms';
-import { DeckCreate } from '../../../../../core/services/deck.service';
+import { DeckCreate } from '../../../../../core/services/api/deck.service';
 
 interface TagCategory {
   name: string;
