+++ frontend/src/app/features/memorize/decks/deck-study/deck-study.component.ts
@@ -3,9 +3,9 @@ import { Component, OnInit } from '@angular/core';
 import { CommonModule } from '@angular/common';
 import { FormsModule } from '@angular/forms';
 import { ActivatedRoute, Router } from '@angular/router';
-import { DeckService } from '../../../../core/services/deck.service';
-import { BibleService } from '../../../../core/services/bible.service';
-import { UserService } from '../../../../core/services/user.service';
+import { DeckService } from '../../../../core/services/api/deck.service';
+import { BibleService } from '../../../../core/services/api/bible.service';
+import { UserService } from '../../../../core/services/api/user.service';
 
 interface StudyVerse {
   verse_id: number;
