+++ frontend/src/app/features/memorize/decks/pages/deck-list-page/deck-list-page.component.ts
@@ -7,8 +7,8 @@ import { DeckListComponent as DeckListPresentationalComponent } from '../../comp
 import { DeckWithCounts } from '../../components/deck-card/deck-card.component';
 import { CreateDeckModalComponent } from '../../components/create-deck-modal/create-deck-modal.component';
 import { DeckFilterComponent } from '../../components/deck-filter/deck-filter.component';
-import { UserService } from '../../../../../core/services/user.service';
-import { ModalService } from '../../../../../core/services/modal.service';
+import { UserService } from '../../../../../core/services/api/user.service';
+import { ModalService } from '../../../../../core/services/utils/modal.service';
 import {
   DeckCreate,
   DeckService,
@@ -16,7 +16,7 @@ import {
   DeckListResponse,
   CardWithVerses,
   DeckCardsResponse,
-} from '../../../../../core/services/deck.service';
+} from '../../../../../core/services/api/deck.service';
 
 interface Tab {
   id: 'my-decks' | 'public' | 'saved';
