+++ frontend/src/app/features/memorize/flow/flow.component.ts
@@ -3,9 +3,9 @@ import { ActivatedRoute } from '@angular/router';
 import { CommonModule } from '@angular/common';
 import { Subject, takeUntil } from 'rxjs';
 import { VerseSelection } from '../../../shared/components/verse-range-picker/verse-range-picker.component';
-import { BibleService } from '../../../core/services/bible.service';
-import { UserService } from '../../../core/services/user.service';
-import { FlowParsingService } from '../../../core/services/flow-parsing.service';
+import { BibleService } from '../../../core/services/api/bible.service';
+import { UserService } from '../../../core/services/api/user.service';
+import { FlowParsingService } from '../../../core/services/utils/flow-parsing.service';
 import { MemorizationModalComponent } from '../../../shared/components/memorization-modal/memorization-modal.component';
 import { User } from '../../../core/models/user';
 import { UserVerseDetail } from '../../../core/models/bible';
