+++ frontend/src/app/features/profile/profile.component.ts
@@ -3,10 +3,10 @@ import { Component, OnInit, OnDestroy } from '@angular/core';
 import { CommonModule } from '@angular/common';
 import { FormsModule } from '@angular/forms';
 import { RouterModule, Router } from '@angular/router';
-import { ModalService } from '../../core/services/modal.service';
+import { ModalService } from '../../core/services/utils/modal.service';
 import { User } from '../../core/models/user';
-import { UserService } from '../../core/services/user.service';
-import { BibleService } from '../../core/services/bible.service';
+import { UserService } from '../../core/services/api/user.service';
+import { BibleService } from '../../core/services/api/bible.service';
 import { HttpClient } from '@angular/common/http';
 import { environment } from '../../../environments/environment';
 import { Subject, takeUntil } from 'rxjs';
