+++ frontend/src/app/shared/components/memorization-modal/components/navigation-controls/navigation-controls.component.ts
@@ -1,8 +1,8 @@
 import { Component, Input, Output, EventEmitter, OnInit, OnDestroy } from '@angular/core';
 import { CommonModule } from '@angular/common';
 import { trigger, transition, style, animate } from '@angular/animations';
-import { RecordingService, RecordingState } from '../../../../../core/services/recording.service';
-import { NotificationService } from '../../../../../core/services/notification.service';
+import { RecordingService, RecordingState } from '../../../../../core/services/utils/recording.service';
+import { NotificationService } from '../../../../../core/services/utils/notification.service';
 import { Subject, takeUntil } from 'rxjs';
 
 @Component({
