+++ frontend/src/app/shared/components/memorization-modal/memorization-modal.component.ts
@@ -9,9 +9,9 @@ import {
 } from '@angular/core';
 import { CommonModule } from '@angular/common';
 import { Router } from '@angular/router';
-import { BibleService } from '../../../core/services/bible.service';
-import { UserService } from '../../../core/services/user.service';
-import { RecordingService } from '../../../core/services/recording.service';
+import { BibleService } from '../../../core/services/api/bible.service';
+import { UserService } from '../../../core/services/api/user.service';
+import { RecordingService } from '../../../core/services/utils/recording.service';
 import { BibleBook } from '../../../core/models/bible';
 import { Subject, takeUntil } from 'rxjs';
 import { trigger, style, transition, animate, keyframes, state } from '@angular/animations';
