+++ frontend/src/app/shared/components/notification/notification.component.ts
@@ -1,7 +1,7 @@
 import { Component, OnInit, OnDestroy } from '@angular/core';
 import { CommonModule } from '@angular/common';
 import { Subscription } from 'rxjs';
-import { NotificationService, NotificationMessage } from '../../../core/services/notification.service';
+import { NotificationService, NotificationMessage } from '../../../core/services/utils/notification.service';
 
 @Component({
   selector: 'app-notification',
