+++ frontend/src/app/state/bible-tracker/effects/bible-memorization.effects.ts
@@ -17,9 +17,9 @@ import {
 } from 'rxjs/operators';
 
 import { BibleMemorizationActions } from '../actions/bible-memorization.actions';
-import { BibleService } from '../../../core/services/bible.service';
-import { UserService } from '../../../core/services/user.service';
-import { NotificationService } from '../../../core/services/notification.service';
+import { BibleService } from '../../../core/services/api/bible.service';
+import { UserService } from '../../../core/services/api/user.service';
+import { NotificationService } from '../../../core/services/utils/notification.service';
 import { 
   selectUserId, 
   selectIncludeApocrypha,
