+++ frontend/src/app/state/bible-tracker/effects/bible-tracker.effects.spec.ts
@@ -10,7 +10,7 @@ import { BibleTrackerEffects } from './bible-tracker.effects';
 import { BibleTrackerActions } from '../actions/bible-tracker.actions';
 import { bibleTrackerReducer } from '../reducers/bible-tracker.reducer';
 import { BibleBook } from '../../../core/models/bible';
-import { BibleService } from '@app/app/core/services/bible.service';
+import { BibleService } from '@app/core/services/api/bible.service';
 
 describe('BibleTrackerEffects', () => {
     let actions$: ReplaySubject<any>;
