+++ frontend/src/app/state/bible-tracker/effects/bible-tracker.effects.ts
@@ -4,7 +4,7 @@ import { Store, Action } from '@ngrx/store';
 import { Observable } from 'rxjs';
 import { map, mergeMap, withLatestFrom, debounceTime, tap } from 'rxjs/operators';
 
-import { BibleService } from '../../../core/services/bible.service';
+import { BibleService } from '../../../core/services/api/bible.service';
 import { BibleTrackerActions } from '../actions/bible-tracker.actions';
 import {
   BibleStatisticsState,
