+++ frontend/src/app/state/practice-session/effects/practice-session.effects.ts
@@ -25,9 +25,9 @@ import {
 } from '../selectors/practice-session.selectors';
 import { BaseEffects } from '../../core/effects/base.effect';
 import { ResponseQuality } from '../models/practice-session.model';
-import { PracticeService } from '@app/app/core/services/practice.service';
-import { AudioService } from '@app/app/core/services/audio.service';
-import { NotificationService } from '@app/app/core/services/notification.service';
+import { PracticeService } from '@app/core/services/api/practice.service';
+import { AudioService } from '@app/core/services/utils/audio.service';
+import { NotificationService } from '@app/core/services/utils/notification.service';
 
 @Injectable()
 export class PracticeSessionEffects extends BaseEffects {
