+++ frontend/tsconfig.json
@@ -20,7 +20,8 @@
     "target": "ES2022",
     "module": "ES2022",
     "paths": {
-      "@app/*": ["./src/*"]
+      "@app/*": ["./src/app/*"],
+      "@environments/*": ["./src/environments/*"]
     }
   },
   "angularCompilerOptions": {
